pyproject.toml
[tool.poetry]
name = "apn-tester"
version = "0.1.0"
description = "test Apple Push Notification with ease"
authors = ["Kelvin Ng <hoishing@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
httpx = { extras = ["http2"], version = "^0.27.0" }
cryptography = "^43.0.0"
pyjwt = "^2.9.0"
streamlit = "^1.37.1"
watchdog = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
